s/int/float

+++ salt/netapi/__init__.py
@@ -258,9 +258,9 @@ class NetapiClient:
         :return: Returns the result from the runner module
         """
         kwargs["fun"] = fun
-        
+
         if timeout is not None:
-            timeout = int(timeout)
+            timeout = float(timeout)
 
         runner = salt.runner.RunnerClient(self.opts)
         return runner.cmd_sync(kwargs, timeout=timeout, full_return=full_return)
